# TestExp/server/common/swagger/Api.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: TestExp
  description: This is for testing generator of express
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
definitions:
  ExampleBody:
    type: object
    title: example
    required:
      - name
    properties:
      name:
        type: string
        description: The example name
      code:
        type: integer
        description: The example code
paths:
...
  /examples/{id}:
    get:
      tags:
        - Examples
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the entity to retrieve
          type: string
      responses:
        200:
          description: Return the example with the specified id
        404:
          description: Example not

    put:
      tags:
        - Examples
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the entity to update
          type: string
        - name: example
          in: body
          required: true
          description: The instance of example entity
          schema:
            $ref: "#/definitions/ExampleBody"
      responses:
        200:
          description: Success to update
        404:
          description: Example not found

    delete:
      tags:
        - Examples
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the entity to delete
          type: string
      responses:
        204:
          description: Success to delete
        404:
          description: Example not found
...